packages/ai/requirements.txt
# MyNaga Gabay AI Service Dependencies
# =====================================

# Core ML
transformers>=4.36.0
torch>=2.0.0

# Audio processing
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing (Prescription Scanner)
pillow>=10.0.0
